deleteTodo delete from id = id AND user_id = userId

+++ 3/src/routes/todoRoutes.js
@@ -34,9 +34,10 @@ router.put('/:id', (req, res) => {
 // Delete a todo
 router.delete('/:id', (req, res) => {
     const { id } = req.params
+    const userId = req.userId
 
-    const deleteTodo = db.prepare('DELETE FROM todos WHERE id = ?')
-    deleteTodo.run(id)
+    const deleteTodo = db.prepare('DELETE FROM todos WHERE id = ? AND user_id = ?')
+    deleteTodo.run(id, userId)
 
     res.json({ message: "Todo deleted" })
 })
